Experimenting with new UX for detected media
Defaulting to the detected image list instead of pre-selecting a
default pict. Needs to be styled on 1 line, scrollable/swipable. Goals
are to make the “widget” more subtle, as well as mixing images and
embeds in said list. Clicking on any would insert it into the TinyMCE
editor, giving preview.

See:
 * #36
 *
https://wordpress.slack.com/archives/feature-pressthis/p1423070672001066