New embed/image detection/presentation/insertion
Continuing with experimenting with a new image/embed presentation and
interaction. Now shows the detected media as a list of small
thumbnails. Embeds are listed 1st and can be clicked to insert embed.
Images can be clicked and inserted as images. List needs to be styled
on one horizontally scrollable/swipable line. We need a generic
thumbnail for embeds when no preview can be deduced from URL (only
possible with Youtube), see `’a bundled thumb representing embed’` in
app.js.