Custom admin screen for PT, for complete control
Adding a new setting screen for PT, so we can style and upgrade the
instal process to our hearts’ content. Also adding a very basic admin
notice to the plugins screen so users are told to go to said admin
screen when PT is activated.

See https://cloudup.com/c5IN6KVcumw