Removing obsolete `get_full_size_src()`.
`get_full_size_src()` is a reminiscence of the initial media selection
UX we tried, but have since discarded. We no longer need to try and get
an image’s maximum quality version, now that we let users simply insert
them into the editor. Since we now only use 64x64px icons to represent
media, we should and now use the actual version of the in-page image,
or a 64x64 version for speed if we can deduct the source supports it.
Added benefit is increased rendering speed.