Issue #67. Adds the classes `is-hidden` and `is-off-screen`.
I replaced some of the old classes to keep consistency (we had `is-open`,
`is-active`, etc).

Some of the declarations are inverted to play well with these new classes.
The idea is not to mix 'is-hidden' and  'is-visible' or 'is-off-screen'
and 'is-on-screen'.

In order to achieve this, the styles were changed:
- All the elements without any of this modifier classes retain their
"visible" and "on screen" state.
- Some of the elements start with the classes `is-hidden` and `is-off-screen`.