The canonical link should always be preferred.
The intent of the canonical link is that it should always be preferred, even over the page's current URL. This is to insure forwarding behavior, clean URLs, to remove campaign tags, and prevent sending traffic to duplicative links. In this context, I think that the canonical link should be preferred over the current URL, as that is the intent of the tag ( see: https://support.google.com/webmasters/answer/139066?hl=en - as well as my own research into the topic at http://hacktext.com/2015/02/peaking-into-facebooks-algorithmic-black-box-2022/ ). This is how Facebook and Google+ both treat the canonical link.